requirements.txt
annotated-types==0.7.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
cymem==2.0.8
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.0
git-filter-repo==2.38.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.6.0
google_labs_html_chunker==0.0.5
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.23.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
parsel==1.9.1
preshed==3.0.9
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
queuelib==1.7.0
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
tldextract==5.1.2
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
Twisted==24.3.0
# twisted-iocpsupport==1.0.4
typer==0.9.4
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
wasabi==1.1.3
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.3
zope.interface==6.4.post2
